Warning fixes:
        * sparc.h (EXTRA_SPECS): Add missing initializers.
        (sparc_defer_case_vector): Provide a prototype.
        * svr4.h (ASM_OUTPUT_ASCII): Cast STRING_LIMIT to (long) when
        comparing it to the result of a pointer subtraction.

From-SVN: r23254

+++ gcc/config/svr4.h
@@ -950,7 +950,7 @@ do {									 \
 	    }								\
 	  for (p = _ascii_bytes; p < limit && *p != '\0'; p++)		\
 	    continue;							\
-	  if (p < limit && (p - _ascii_bytes) <= STRING_LIMIT)		\
+	  if (p < limit && (p - _ascii_bytes) <= (long)STRING_LIMIT)	\
 	    {								\
 	      if (bytes_in_chunk > 0)					\
 		{							\
